fix posts.user_id column type

+++ db/migrate/20240708191938_remove_not_null_user_id_from_posts.rb
@@ -1,5 +1,5 @@
 class RemoveNotNullUserIdFromPosts < ActiveRecord::Migration[6.1]
   def change
-    change_column :posts, :user_id, :integer, null: true
+    change_column :posts, :user_id, :bigint, null: true
   end
 end
